Add firebase function to update agreement active contributors
To simplify querying during CLA verification, this patch defines a
firebase function that updates agreements with a set of active
contributor identifiers (email, github IDs). The function is triggered
by creation of new addendums. If a user is added by an addendum, the
corresponding identifiers will be added to the agreements. Otherwise, if
the user is in the removed section of the addendum, identifiers will
be removed from the agreement.