Update whitelist using new identity model
Instead of updating the parent agreement we maintain a new collection
called 'whitelists'. It seems more appropriate since the agreement is an
object that should not be modified once created.

Also, this patch disables the unknown identity for now (not clear what's
the use case) Contributor functions have been moved to to Cla lib.