Rework github webhook to handle requests asynchronously
The payload of a PR webook is now stored in the DB and processed
asynchronously. This allows us to keep track of PRs that are not covered
by a CLA. When a new addendum is added and the whitelist updated, we can
scan for all pending requests and process them right away, without any
extra actions from users (TODO).