Rework github webhook to handle events asynchronously (#45)
* Rework github webhook

PR webooks are now stored in the DB and processed asynchronously. This
allows us to keep track of PRs that are not covered by a CLA. When a new
addendum is added and the whitelist updated, we can scan for all pending
requests and process them right away, without any acctional action from
users. That has yet to be implemented yet, but the database schema is in
place.

Two new collections are maintained: (1) contributions, and (2) events.
One contribution is created for each PR, while an event is created for
each PR webhook action (open, synchronize, close, reopen).

When processing events, the implementation checks the whitelists for the
PR author and updates the PR status, eventually posting a comments in
case user action is needed, e.g. asking to sing a CLA, or for errors.

* Fix teo's comments

* Fix tests not passing

* Use warn instead of error